+++ .gitignore
@@ -266,3 +266,4 @@ __pycache__/
 
 # Cake - Uncomment if you are using it
 # tools/
+/Source/Nuget
